Data types supporting date ranges outside of those supported by System.DateTime and the greater variety and detail provided by postgres' date and time datatypes.
Moved Npgsql into the same file, as it is similar in some ways.

Becaues these datatypes can clash with cases where one uses System.DateTime with a DataAdaptor, added a connection string option to choose between the the two.

Also added error handling for the case where System.DateTime is being used, and an unexpected value outside of its range appears in data.