Conn can be closed in TransactionScope without prepared transaction
Previously closing a connection within a TransactionScope triggered an
automatic promotion to a prepared transaction. Since by default Postgres
disables these, this made the TransactionScope unusable for many cases.
This is reproduced in
SystemTransactionTests.SimpleTransactionScopeWithImplicitEnlist, which was
failing in stock Postgres installations.

Instead of promoting, we now postpone the close until the (local)
transaction is terminated, i.e. until the TransactionScope ends.