Add support for new error fields added in PostgreSQL 9.3
This makes the new error information available in the Exception class
which allows the calling code to, for example see which constraint would
be violated by the request.
See PostgreSQL commit message for more details on what extra information
is available and when. This commit only captures the extra information
when it's there.

http://git.postgresql.org/gitweb/?p=postgresql.git;a=commit;h=991f3e5ab3f8196d18d5b313c81a5f744f3baaea