Adds new test for newly added exception fields
This code is untested as I've not yet worked out how to run the test
suit.
The constant values in the Assert.AreEqual() calls may need tweaked.
The prime purpose of the patch is to test that the fields are properly
being populated. There is no need to also test things like foreign key
violation as its testing the same code in Npgsql.