Fixed wrong assumptions about new Exception Tests
After reading the source code of PostgreSQL9.3, it seem that only
certain errors relating to domain types will populate the DataTypeName
field. This commit adds 2 new tests. 1 to test for the column name, as
the previous uniqueviolation did not show the column name which unique
constraint was violated. Perhaps this has to do with the fact that
unique constraints can exist for >1 column. Test 2 checks the
datatypename after a domain's NOT NULL constraint is violated.

I've not yet run the tests. Though I did compare after setting psql \set
VERBOSITY verbose