Tidy up of new test functions
Renamed the 3 new test functions to give them more suitable names. Also
added comment to explain why a domain is created to test the
DataTypeName field. I would have guessed that we could see this
populated by violating a NOT NULL constraint on a table's column, but
this does not seem to be the case.