Replaced SslClientStream with .NET SslStream
This change fixes an issue when connecting to a PostgreSQL database
configured to accept connections only with a valid client certificate.
With Mono's SslClientStream, an encryption error would occur when trying
to connect. WIth .NET's SslStream, the connection can be successfully
opened. The Mono-based callbacks have been marked "Obsolete". The newer
ValidateRemoteCertificateCallback is used automatically if it is
provided.