Binary test updates
Add internal flag that fully suppresses binary backend parameter and result formatting.
Add reflection-based code to set and observer this flag in the test suite.
Add a test that reports failure to bind to the flag.
Add a couple tests that suppress binary backend encoding to make sure text encoding is tested on types that support binasy encoding.