Fix extra_float_digit setup for Postgresql below 9.0
Npgsql was sending two commands at the same time to set the extra_float_digits parameter.
The first one with value of 2 and the second one with a value of 3.
The problem is that on postgresql below 9.0, there is no support for value of 3 and then neither
command was being processed letting the parameter with the default value of 0.
Now, Npgsql checks which version it is connected to and sends the correct command accordingly.