Auto-build AsyncGenerator from Npgsql if doesn't exist
Although it's required for the build process, the AsyncGenerator project
cannot be regular dependency of npgsql, because it would get distributed
with Npgsql. However, we want a AsyncGenerator to get build if necessary,
so that the build process can remain simple (e.g. fresh clones work).

Added a check in the Npgsql project to see if AsyncGenerator has been
built, and if not, builds it.

Warning: this will not rebuild if AsyncGenerator is out of date, only if
it hasn't been built at all.