Fixed bug when reading large values in sequential mode
Renamed IsArbitraryLength property to CanReadFromSocket on type handlers.
For handlers which *don't* know how to read from the socket, we previously
assumed their column length would always be smaller than the buffer. This
isn't true since we have some arbitrary-length types for which it's not worth
the effort to implement socket-reading behavior (e.g. path).

So we now call EnsureOrAllocateTemp() instead - in this rare worst case a
temporary buffer will be allocated to hold the value.

Thanks @Emill