Remove the NpgsqlConnector.IsValid method.
Npgsql no longer checks if the connector is valid avoiding roundtrips to the server just to check that.
IsValid also run the initQueries used to reset the Connector state using user configurations. This reset
was moved to a Reset() method. Now NpgsqlConnectorPool resets the connector using this method
before returning it from the pool.

Conflicts:
	Npgsql/NpgsqlConnector.cs