Command no longer subscribes to connection's state event
NpgsqlCommand was listening to its connection's StateChange event so as
to know when the latter was closed, in order to clear its prepared status.
However, it seems that Entity Framework does not dispose of commands, which
led to a leak (since commands never get unsubscribed from their connections).

Rewrote the mechanism without events: a connection maintains a counter that
gets incremented on each open, and the command knows for which counter (i.e.
connection instance) it was prepared.

Backport of 31bdccec5fbf8f6020c2f78ba1a3c50d592e06e5

Fixes #505