Command no longer subscribes to connection's state event
NpgsqlCommand was listening to its connection's StateChange event so as
to know when the latter was closed, in order to clear its prepared status.
However, it seems that Entity Framework does not dispose of commands, which
led to a leak (since commands never get unsubscribed from their connections).

Rewrote the mechanism without events: connection knows when it was last closed,
command knows when it was prepared, check the times

Fixes #505