Command no longer subscribes to connection's state event
NpgsqlCommand was listening to its connection's StateChange event so as
to know when the latter was closed, in order to clear its prepared status.
However, it seems that Entity Framework does not dispose of commands, which
led to a leak (since commands never get unsubscribed from their connections).

Rewrote the mechanism without events: a connection maintains a counter that
gets incremented on each open, and the command knows for which counter (i.e.
connection instance) it was prepared.

Fixes #505