Better cleanup of idle connectors on shutdown
When the process or appdomain is being shut down, we want to
send Terminate messages to the backend on idle connectors in
the pool for a clean shutdown. We were doing this via a finalizer,
but this is a dangerous mechanism (not supposed to assume referenced
objects are still there).

Switched to listening to the AppDomain.DomainUnload and ProcessExit
events instead.

Also did better handling of connection errorsbetter handling of connection
errors

Fixes #491