Added krbsrvname parameter and use hostname for SSPI authentication
The GSS and SSPI authentication needs the kerberos service name of the
PostgreSQL server. The default value is POSTGRES but this value can be
changed. The ODBC driver allows to set this value using the
configuration option "krbsrvname". This patch allows to send this
parameter in the connection string.

If the ip address is used instead of the hostname, the SSPI
authentication process uses NTLM authentication. By using the host name
it tries first to use Kerberos and if this is not possible NTLM is used.