Type mapping DbType inference fix
We had case like json, varchar (and other text aliases) not covered by the type
mapping system. If a parameter's DbType is set to json, we want Npgsql to map
to PG text. However, when these types are explicitly selected with
NpgsqlDbType.Json or NpgsqlDbType.Varchar, we still need the parameter's DbType
to be set to DbType.String.

In addition, in cases where multiple DbTypes where mapped to a single type
(e.g. String, StringFixedLength, AnsiString...), it was unclear which DbType
should be the "primary" one, i.e. the one the parameter's DbType is set to.

Added an InferredDbType property on the TypeMapping attribute to handle this.