Refactor async notifications and message draining
Previously, any code that wanted to communicate with the server has to use
a NotificationBlock for synchronization with the async notification handling
mechanism.

Refactored this so that setting the connector's state to Executing/Fetching/
Copy acquires the lock, and setting it back to ready releases it. This
concentrates the synchronization logic in one place.

Refactored some of the async notification mechanism as well.

Fixes #578