Set statement_timeout in startup packet
Rather than set that backend timeout on the first command, set it in the
packet. It's a bit cleaner and saves resetting the timeout after a
pooled connection close.

Also somewhat important for pgbouncer in transaction mode, which is
incompatible with statement timeouts. When we attempt to set
statement_timeout is set in the startup packet a clear failure message
is provided. When it is set later pgbouncer appears to silently drop
the connection.

Fixes #616