Remove spurious warning while cancelling
After we send a cancellation, PostgreSQL immediately closes the connection.
We wait for this by reading the network stream, but in some cases this
produces an exception.

Added a check for "connection closed by peer" to not log a warning.

(cherry picked from commit 2eccf060bf003040916c16e0d5c39ec108e085af)