Fix connection close with open reader
We previously had an optimization whereby an open reader at connection close
time was closed asynchronously to return control to the user earlier. This
seems to be incompatible with ASP.NET, which doesn't allow async in
arbitrary places.

Removed the optimization, if an open reader exists when closing a connection,
it will be closed synchronously.

Also refactored and fixed closing behavior for non-pooled connectors.

Fixes #666