Fixed close with failed transaction and custom timeout
When closing a pooled connection with a transaction, a rollback is sent.
Since ROLLBACK is an internal timeout, if the current backend timeout
is is different from the internal timeout (e.g. some custom command was
set), a "set timeout" is sent before the rollback.

However, if we're in a *failed* transaction state, the "set timeout" fails.
Added a check for the failed transaction state, skipping the "set timeout"
(this was already the case for a manual rollback, but not for an automatic
rollback on connection close).

Fixes #719

(cherry picked from commit 638c54ca717bf217dddc3613814d05656cdab9a1)