Fix overflow when opening/closing same connection many times
Connector has a private _pendingRfqPrependedMessages which tracks the number
of currently prepended messages which expect an RFQ response. By simply
opening and closing the connection many times, it was possible to overflow
this field (each connector reset prepends a DISCARD ALL).

We now clear any pending messages when resetting the connector.

Fixes #736