Fix transaction state bug
If a timeout and a rollback were already enqueued (because a connection with
a transaction was returned to the pool and reset), the rollback response's
Idle transaction status was processed and caused new transactions to get
cleared.

We now process transaction status only on non-prepended messages.

Fixes #765

(cherry picked from commit 2ae061524edc153e72315cb789422f51daa1edd6)