Take into account CopyData header when importing
TextHandler's Write logic looks at the buffer's size as part of its decision-
making. However, during import the buffer always has a CopyData header (5
bytes), and this caused us to enter an endless loop when trying to write
strings with length 8192 or just under.

Introduced UsableSize on the buffer to account for this.

Fixes #816