[EF7] Ensure schema with CREATE SCHEMA IF NOT EXISTS
Previous method attempted to work around the lack of CREATE SCHEMA IF NOT
EXISTS in PostgreSQL < 9.3 by creating a function which checks and creates
the schema, and then invoking it. This failed because of #641 (pg_temp
doesn't exist yet).

Switch back to simple CREATE SCHEMA IF NOT EXISTS which will fail with
PG < 9.3. Skip schema creation for public which is a workable workaround.

Fixes #838