Try to fix bug appearing on Npgsql Teamcity
For some reason the Npgsql Teamcity build fails telling "Cannot set
NpgsqlDbType to just Array, Binary-Or with the element type (e.g. Array
of Box is NpgsqlDbType.Array | NpgsqlDbType.Box)" which dosn't happen
on my development system.
This is an attempt to fix this without really knowing what's going on.