Remove deferred rollback-on-close optimization
When a pooled connection was closed while a transaction was still in
progress, Npgsql had an optimization whereby it didn't execute a rollback
immediately, but rather prepended a rollback message for execution the
next time the connection is opened and used. This created some issues as
the transaction kept locks etc. for an indefinite time.

Removed the optimization, closing a connection now immediately sends a
rollback.

Fixes #927