Do not restart the timer if it's already running
Previously we restarted the timer every time a connection was
requested. If a server (not database) had a connection request per
second or more then the timer would never actually fire and the pool
will never get cleaned up. The solution then is to just not fire the
timer if it's already running.