NextResult() now throws statement exceptions
In multistatement commands, if a later statement generated an error the
exception would be thrown on the first Read() of the statement's
resultset.

Changed the logic to have NextResult() always read the first message,
throwing the exception at that point if one exists. This also simplified
the reader logic.

Fixes #1032