Pool checks for broken connections
The keepalive mechanism can cause a connection to break while it's in the
pool, the pool should check if a connection is in the ready state before
returning it. Because of #1048 this wasn't properly tested and did not
work.

Fixed #1050

(cherry picked from commit c9ecc548ea9d39cb977ada47d3ef1f9d5bbbf618)