Don't send ROLLBACK on close when tx pending
When reseting a connection as part of a pooled connection close, we used
to send ROLLBACK to the server even if the transaction was only pending,
i.e. in our write buffer but not sent to the server. This generated a
warning in the PG logs.

Fixes #1203

(cherry picked from commit 3667ec34126c717b4f4017087be1812b8f7c2f85)