Lazily compute NpgsqlConnection.ConnectionString
To make connection open faster, the string ConnectionString property is
now computed lazily - it was previously set when opening (including
password removal when Persist Security Info is false, etc.).

It's now possible to only interact with the connection using the strongly-
typed NpgsqlConnectionStringBuilder and never waste time on parsing or
reconstructing the ConnectionString string.

Fixes #1174