String writing is more efficient
We now use the pointer overload of Encoder.Convert, removing an allocation
and copy of many strings written to the backend. Affects both query SQL
(which was always copied) and big string parameters which don't fit in the
buffer.

Unfortunately this optimization isn't possible in netstandard13 due to the
lack of the pointer overload.

Closes #1233