Fix pool release deadlock
When under load and opening attempts are pending, a connection close
hands off the connector directly to the enqueued opening attempts. Part
of this handoff occurred in the TP, causing a deadlock when the thread
pool was exhausted with synchronous open attempts (which were enqueued).

We now track whether open attempts are sync or async,
and only turn to the TP during handoff if the attempt is async.

Fixes #1232

(cherry picked from commit 85874449c6f41363dc3ba71e48ba55a62071790f)