Fix multistatement deadlock
When sending large multistatement commands, we send "remaining" query data
asynchronously, even if the user is using a sync API.
This remaining async send used to happen on the TP, triggering a
deadlock if the user's sync call was on the TP as well.

Remaining async sends now run with a special synchronization context which
executes continuations on a special dedicated thread, avoiding the thread
pool.

All participating async methods have been duplicated to remove
ConfigureAwait(false) in order to preserve the synchronization context
(but see #1282 for a better solution for 3.2).

Fixes #1256