Only fetch rich column metadata if KeyInfo is set
When calling NpgsqlDataReader.GetColumnSchema (or
NpgsqlDataReader.GetSchemaTable which has been reimplemented on top of
it), additional column metadata can be fetched from the database via an
additional query. However, this should only occur if the command's
CommandBehavior contains KeyInfo. This is important for the performance of
DataTable, which calls GetSchemaTable without KeyInfo.

Fixes #1346, fixes #1343