Fix issue resetting send state
An executing command instance is notified by WriteBuffer if the ongoing
send process flushed the buffer - this is necessary to ensure the
switch to async sending to avoid deadlocks in multistatement commands.

At the end of the send, the command's CleanupSend reset the
WriteBuffer's reference to the command. However, once execution has
started, the command's state can change (e.g. its connection can be
set to null), so this was unsafe and could cause exceptions in
race conditions.

No longer reset the reference, it will simply be overwritten the next
time a send starts.

Fixes #1398