Fix pooling issue returning bad connection
If the same instance of NpgsqlConnectionStringBuilder is reused to open
two connections, but a property is modified on that instance, the pool
erroneously returned the same physical connection.

Switch from ConcurrentDictionary to regular Dictionary with lock, and
clone the connection string builder instance before inserting it as
a key.

Fixes #1412