Fix column schema for non-columns
When getting the schema for non-columns, nullability is unknown since
we only have the RowDescription, and can't query an actual table schema
in PostgreSQL. As a result, AllowDBNull is set to null.

However, when using the old schema API
(NpgsqlDataReader.GetSchemaTable()), Npgsql mistakenly returned false
for non-columns, making DataTable fail.

Fixes #1497