Preparations for targeting .NET Standard 2.0
Inversed all #ifdefs to consider netstandard13 and netcoreapp11
as exceptions since netstandard20 covers almost all of the .NET
Framework API surface. In some rare cases it's still necessary
to check specifically for net45/net451 (e.g. performance
counters, directory services).

It's enough to add netstandard20 and netcoreapp20 to Npgsql and
to Npgsql.Tests respectively to support .NET Standard 2.0 (this
is currently not done because not supported at Appveyor/Travis).

Relates to #1575