Cancel COPY import if closing in middle of row
If a COPY importer was closed in the middle of a row, an exception was
thrown. This broke the connection, and also masked the original
exception by throwing a new one.

Instead of throwing an exception from Dispose(), we simply cancel the
current import.

See #1646 for an additional related API change.

Fixes #1645

(cherry picked from commit 90a7aeaa31c51e31af8125b6954ac015d71e6e87)