handle CrateDB server version psql parameter
CrateDB vesions >= 2.3 sends the current server version as a parameter
over the Postgres Wire Protocol. This is useful to adopt the server
capabilities and it makes the DB detection for 3rd party tools easier
as they might might not be able to configure the connection string using
the "Server Compatibility Mode" explicitly.