Optimize datarow field processing
In non-sequential reading mode (the default), we were previously
traversing the DataRow twice - once to construct a column
offset table, and once when the reader actually consumes the columns.
Although the first traversal was relatively efficient (simple memory
seeks), made the mechanism lazy so that column offsets are
discovered as part of user column consumption.

Also optimized row consumption (seek to end).