Reimplemented pooling as lock-free
ConnectorPool is now fully lock-free - benchmarks showed consierable
contention on the lock in the previous implementation in highly
parallel scenarios.

Also reimplemented PoolManager. The previous implementation used a
ConcurrentDictionary, but 99% of all apps will have a single pool and
the lookup is needlessly expensive. It is now an append-only
array of tuples where reading is a lock-free operation (but not adding).

Note that this is needed only because ADO.NET doesn't have a
proper user-facing API which exposes pooling to the user, otherwise
the user could simply take a reference to the pool and avoid this
(complex) lookup altogether. See:
https://github.com/dotnet/corefx/issues/26714

Also corrected some inaccurate performance counter updating.

Closes #1839