Remove connector semaphore/lock if no keepalive
Whenever a user action was started, we previously took a SemaphoreSlim,
which was itself guarded by an ordinary lock (since
SemaphoreSlim.Dispose() isn't threadsafe). This served two purposes:
to throw a meaningfull exception to users in case of concurrent usage,
and to synchronize when keepalive is enabled.

If keepalive isn't enabled, we no longer take the semaphore or lock,
but instead rely Interlocked state transition to detect concurrent
use on a best-effort basis.