Fix pool state bug on exception when type loading
If an exception is thrown during type loading, the pool's busy counter
got incremented twice: once from NpgsqlConnector.Break() (since the
connector is no longer in state Connecting), and once in the pool
try/catch code calling the Open().

Added special state tracking whether we're connecting, checking for
that in Break() and avoiding calling release on the pool.

https://github.com/npgsql/npgsql/commit/45e33ecef21f75f51a625c7b919a50da3ed8e920#r28235021

Thanks @FlorianRainer