Pooled connection dynamic stats collection
Context: https://github.com/npgsql/npgsql/issues/1914#issuecomment-388638067

Previously, when `NpgsqlReadBuffer` was constructed it would check to see if the user wanted to collect stats. And because `NpgsqlReadBuffer` is only constructed once per connection, and if that connection is then later on pooled that would meanly that we would *always* collect stats even if the user disabled stats for a given session. Conversely, if the pooled connection was constructed with stats disabled, stats collection would *always* be disabled regardless if the user enabled stats collection for a given session.

Now, we no longer check to see if the user wants to collect stats at the time of construction. But right before we lose the information we are interested in. Unfortunately we incur the cost of a two null checks and bool check. The null checks are required because the `NpgsqlConnector` can be null, and the `NpgsqlConnection` inside of the `NpgsqlConnector` can be null too.