Adds compatibility flag for enum loading behavior
Enums are loaded for version 8.3 and later, but the pg_enum
only has the 'enumsortorder' column beginning in version 9.1.

To maintain compatibility, this commit adds a check for the later
version, and omits the sort order when unavailable.

Version 8.3 is past its end-of-life, but its cheap to add a
compatibility flag, so it seems worthwhile.

This commit also adds some doc comments and nullability attributes
to `PostgresDatabaseInfo` and `PostgresDatabaseInfoFactory`.

(cherry picked from commit a74359a939c18efd8aad75d66f85388f0939b3d9)