Bad hack to reallow output params with sequential
Npgsql 4.0 accidentally removed support for output parameters when
using CommandBehavior.Sequential, but unfortunately DbDataAdapter
relies on that.

4.0's refactoring into two reader classes (sequential, non-sequential)
made this support very hacky, as the 1st row needs to be traversed
twice, once for output param population and once as a regular row.
We definitely need to clean this up in 4.1.

Fixes #2091