Don't throw from NpgsqlFactory.GetService()
The Entity Framework 5 way of getting its service provider is via
NpgsqlFactory.GetService(). If this method was called and the EF5
assembly can't be found, we used to throw. This is problematic in
case where Visual Studio attempts to load EF5 in case it is
installed - we now return null instead.

Fixes #2271