Handle exception when timezone is `localtime` (#2264)
The configuration property `timezone` in `postgresql.conf` can be
set to a special value of `localtime` to indicate that PosgreSQL
should use the timezone of the host system.

Unfortunately, when this value is set, we do not have sufficient
information to determine the actual timezone of the server because
`SHOW timezone` itself returns `localtime`.

This value is not well-documented, but is also unlikely to be
encountered very often, so we identify the exception and construct
a new `TimeZoneNotFoundException` with a more informative
message that includes mitigations that can be applied on either
the server or the client.

See: npgsql/Npgsql.EntityFrameworkCore.PostgreSQL#578