Use Span for Substring in Netstandard 2.1+
If we can append a Span to a StringBuilder (Netstandard 2.1+) we will do
that instead of using string.Substring in order to reduce allocations.
Note that in two spots we must convert the Span back to a string, once
so that we can use it in a dictionary and the other in order to use a
string only version of IndexOf.