Add unit of measure and default to ConnectionTimeout
The summary, not value, is what is shown by consumers  when using autocomplete.  This change allows consumers to easily see the unit of measure (seconds) as well as the default value when using autocomplete/code completion

I'm proposing this change because it wasn't clear to me if the timeout was in seconds or milliseconds.