Removed AwaitableSocket
The class was made in time when async-await was poor and not each method in .NET was optimized to be used with it, but now .NET is mature and NetworkStream calls async methods on a Socket which internally uses cached SocketAsyncEventArgs.

Fixes #2807.

(cherry picked from commit baa343475ce1bbc19b1d2893386aeeab1c87a6df)