Add NpgsqlLogSequenceNumber and PgLsnHandler to handle the pg_lsn type
This type was added in PostgreSQL 9.4 and we didn't handle it yet.
While a type handler and a specific type isnt't strictly necessary for
the implementation of replication in Npgsql it will probably be useful
once people start using the replication feature and want to query their
database regading replication related information like
pg_replication_slots.