Adjust replication API for PostgreSQL 14
PostgreSQL 14 brings several changes to the Logical Streaming
Replication Protocol. This commit brings support for the following:

1. Logical decoding messages
   These are messages you can use to send text and arbitrary binary
   data to logical decoding plugins (and now also to a client through
   a replication connection).
2. Streaming of large in-progress transactions
   With this feature the server can start streaming large transactions
   to the client while it is still processing them

The second of those features (streaming transactions) is supported via
the Logical Streaming Replication Protocol version 2 which  which is now
supported for all message types.

A third PostgresQL 14 feature (processing logical replication data in
binary format) will follow in a separate commit.

Since this is the first major revision of the Npgsql logical replication
API which has been considered somewhat instable for now, we allow
ourselves to also introduce a few breaking changes for Npgsql 6.

As first breaking change, this commit changes RelationMessage.Columns
and TruncateMessage.RelationIds to use ImmutableArray<T>.

Due to an oversight in the initial implementation
TruncateMessage.RelationIds was mutable. Also for some reason @roji and
I cannot remember we chose ReadOnlyMemory<T> for the immutable
collections which doesent look like the best or most obvious chice in
hindsight.
Since we'll introduce a buch of breaking changes in the
Npgsql.Replication.PgOutput namespace anyway, I took the opportunity to
change this for RelationMessage.Columns.
I didn't bother to change it for the APIs that contain actual payload
data (e. g. InsertMessage.NewRow) since those APIs will most likely be
converted to a IDataRecord or something like that in a follow-up PR to
enable replication of binary values.