Conditionally ignore two tests that fail on a self-built PostgreSQL
When preparing my system to write tests for the replication PR I noticed
that the test suite fails on my self-built PostgreSQL instance with
default settings because it is lacking features (xml) or has them
disabled by default (prepared transations).
This commit makes us ignore the tests in case of excepions that come
from the fact that the features are missing.