Fix timestamp in trace logging message
Since we trace log the local timestamp when we receive a primary
keepalive streaming replication message we should also do so when
we send a keepalive. Otherwise the timestamps that appear next to
eachother in the trace log will be offset in non GMT time zones.

Also fix a minor whitespace.