Add script to create and update catalogs.json
Add a Perl script to generate a json file (catalogs.json)
containing information about hard-coded types in PostgreSQL,
which we get from the catalog files in the PostgreSQL source tree.
The script will also check whether catalogs.json is up to date and
will greate a pull request with changes via the GitHub API if it is
not.