Make NpgsqlConnector.Transaction setter and UnboundTransaction private
Narrow down the internal interface of NpgsqlConnector by making the
setter of the NpgsqlConnector.Transaction property private and moving
the NpgsqlConnector.UnboundTransaction property to a private field.

Instead of the externally settable properties add two new methods
to NpgsqlConnector:

1. InitializeTransaction()
   Initializes the Transaction property, starts a new transaction
   and returns it.
2. UnbindTransaction()
   Unbinds any undisposed transaction from the Transaction property
   and handles the potential recycling of the latest unbound
   transaction.

The changes above also make most of NpgsqlTransaction.UnbindIfNecessary()
obsolete and the remainder has been replaced with two tiny methods:

1. ReleaseConnector()
   Releases the internal connector by setting it to null to avoid
   concurreny issues.
2. ResetConnector()
   Sets the internal connector to a new connector for recycling
   purposes.

While this commit is mostly refactoring it also adds one minor
behavior change in NpgsqlTransaction:

While the async DisposeAsyncInternal() method was logging exceptions
that happended during disposal, this wasn't the case for the sync
Dispose() method. This commit also adds logging of exceptions that
occur in the sync code path.