Fix connect timeout when resolving to multiple IPs
When the host resolves to multiple IP endpoints in NpgsqlConnector.ConnectAsync and a timeout is used, the timeout was not shared among IPs as intended. Instead, effectively the total timeout was divided by the number of endpoints, and if the first attempt failed after said timeout, the second one would immediately result in a hard timeout error.

The logic was seemingly copied from `void Connect()` where the approach is in fact correct because the `perEndpointTimeout` was relative, whereas in `ConnectAsync()` it was an absolute timestamp due to conversion to `NpgsqlTimeout`.